Fix success message rounding error

+++ prototype/src/afk/transactions/Deposit.java
@@ -28,7 +28,7 @@ public class Deposit extends Transaction {
     }
 
     public String getSuccessMessage() {
-        return "Successfully withdrew $" + (getAmount()/100)
+        return "Successfully withdrew $" + ((double) getAmount()/100)
                 + " from account #" + getSourceAccount().getNumber();
     }
 }
